+++ frontend/src/pages/SessionPage.tsx
@@ -2,7 +2,7 @@ import VideoContainer from "@/components/session/VideoContainer.tsx";
 import { useParams } from "react-router-dom";
 import SessionSidebar from "@/components/session/SessionSidebar.tsx";
 import SessionToolbar from "@/components/session/SessionToolbar.tsx";
-import { useSession } from "@/hooks/useSession";
+import { useSession } from "@/hooks/session/useSession";
 import useSocket from "@/hooks/useSocket";
 import SessionHeader from "@/components/session/SessionHeader";
