+++ frontend/tsconfig.app.json
@@ -1,6 +1,9 @@
 {
   "compilerOptions": {
-    "typeRoots": ["src/types", "node_modules/@types"],
+    "typeRoots": [
+      "src/types",
+      "node_modules/@types"
+    ],
     "tsBuildInfoFile": "./node_modules/.tmp/tsconfig.app.tsbuildinfo",
     "target": "ES2020",
     "useDefineForClassFields": true,
@@ -40,10 +43,12 @@
       ]
     },
     "types": [
-      "node", "vite/client"
+      "node",
+      "vite/client"
     ]
   },
   "include": [
-    "src", "vite-env.d.ts"
+    "src",
+    "vite-env.d.ts"
   ]
 }
