+++ frontend/tsconfig.test.json
@@ -11,7 +11,16 @@
     "moduleResolution": "node",
     "paths": {
       "@/*": [
-        "./src/*"
+        "src/*"
+      ],
+      "@components/*": [
+        "src/components/*"
+      ],
+      "@hooks/*": [
+        "src/hooks/*"
+      ],
+      "@stores/*": [
+        "src/stores/*"
       ]
     }
   },
